requirements.txt
Flask>=3.0.0
ultralytics>=8.0.196
opencv-python>=4.8.1.78
numpy>=2.1.0
Pillow>=10.1.0
torch>=2.1.0
torchvision>=0.16.0
